+++ src/main/java/com/example/simplesns/common/config/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package com.example.simplesns.config;
+package com.example.simplesns.common.config;
 
 import at.favre.lib.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Component;
