+++ src/main/java/com/example/simplesns/common/consts/Const.java
@@ -1,4 +1,4 @@
-package com.example.simplesns.common;
+package com.example.simplesns.common.consts;
 
 public class Const {
     public static final String LOGIN_USER = "loginUser";
