+++ src/main/java/com/example/simplesns/common/filter/LoginFilter.java
@@ -1,6 +1,6 @@
-package com.example.simplesns.filter;
+package com.example.simplesns.common.filter;
 
-import com.example.simplesns.common.Const;
+import com.example.simplesns.common.consts.Const;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,7 +8,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.util.PatternMatchUtils;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
